Refine starter flow template structure
